refactor: fix imports gf-557
