+++ apps/backend/src/libs/modules/server-application/libs/types/types.ts
@@ -1,3 +1,3 @@
-export { type ServerApplicationApi } from "../types/server-application-api.type.js";
 export { type ServerApplication } from "./server-application.type.js";
+export { type ServerApplicationApi } from "./server-application-api.type.js";
 export { type ServerApplicationRouteParameters } from "./server-application-route-parameters.type.js";
