+++ packages/shared/src/modules/project-groups/libs/types/types.ts
@@ -1,6 +1,6 @@
-export { type ProjectGroupCreateRequestDto } from "../types/project-group-create-request-dto.type.js";
-export { type ProjectGroupCreateResponseDto } from "../types/project-group-create-response-dto.type.js";
-export { type ProjectGroupGetAllItemResponseDto } from "../types/project-group-get-all-item-response-dto.type.js";
+export { type ProjectGroupCreateRequestDto } from "./project-group-create-request-dto.type.js";
+export { type ProjectGroupCreateResponseDto } from "./project-group-create-response-dto.type.js";
+export { type ProjectGroupGetAllItemResponseDto } from "./project-group-get-all-item-response-dto.type.js";
 export { type ProjectGroupGetAllRequestDto } from "./project-group-get-all-request-dto.type.js";
 export { type ProjectGroupGetAllResponseDto } from "./project-group-get-all-response-dto.type.js";
 export { type ProjectGroupPatchRequestDto } from "./project-group-patch-request-dto.type.js";
